Add apictl code changes to add API
