+++ import-export-cli/box/resources/kubernetes_resources/api_cr.yaml
@@ -1,11 +1,6 @@
-apiVersion: wso2.com/v1alpha1
+apiVersion: wso2.com/v1alpha2
 kind: API
 metadata:
   name:
 spec:
-  definition:
-    swaggerconfigmapNames:
-    type: swagger
-  replicas:
-  mode: privateJet
-  override: false
+  swaggerConfigMapName:
