+++ import-export-cli/cmd/deleteAPI.go
@@ -56,7 +56,7 @@ var DeleteAPICmd = &cobra.Command{
 		if configVars.Config.KubernetesMode {
 			k8sArgs := []string{k8sUtils.K8sDelete, k8sUtils.ApiOpCrdApi}
 			k8sArgs = append(k8sArgs, args...)
-			executeKubernetes(k8sArgs...)
+			ExecuteKubernetes(k8sArgs...)
 		} else {
 			cred, err := GetCredentials(deleteAPIEnvironment)
 			if err != nil {
