+++ import-export-cli/cmd/deprecated/installWso2amOperator.go
@@ -20,8 +20,8 @@ package deprecated
 
 import (
 	"fmt"
+	"github.com/wso2/product-apim-tooling/import-export-cli/cmd/k8s"
 
-	"github.com/wso2/product-apim-tooling/import-export-cli/cmd"
 	k8sUtils "github.com/wso2/product-apim-tooling/import-export-cli/operator/utils"
 	"github.com/wso2/product-apim-tooling/import-export-cli/utils"
 
@@ -44,7 +44,7 @@ var installWso2amOperatorCmdDeprecated = &cobra.Command{
 	Short:   installWso2amOperatorCmdShortDesc,
 	Long:    installWso2amOperatorCmdLongDesc,
 	Example: installWso2amOperatorCmdExamples,
-	Deprecated: "use \"" + cmd.K8sCmdLiteral + " " + cmd.K8sInstallCmdLiteral + " " + cmd.K8sInstallWso2amOperatorCmdLiteral +
+	Deprecated: "use \"" + k8s.K8sCmdLiteral + " " + k8s.K8sInstallCmdLiteral + " " + k8s.K8sInstallWso2amOperatorCmdLiteral +
 		"\" " + "instead of \"" + installCmdLiteral + " " + installWso2amOperatorCmdLiteral + "\".",
 	Run: func(cmd *cobra.Command, args []string) {
 		utils.Logln(fmt.Sprintf("%s%s %s called", utils.LogPrefixInfo, installCmdLiteral, installWso2amOperatorCmdLiteral))
