+++ import-export-cli/cmd/deprecated/uninstall.go
@@ -21,6 +21,7 @@ package deprecated
 import (
 	"github.com/spf13/cobra"
 	"github.com/wso2/product-apim-tooling/import-export-cli/cmd"
+	"github.com/wso2/product-apim-tooling/import-export-cli/cmd/k8s"
 	"github.com/wso2/product-apim-tooling/import-export-cli/utils"
 )
 
@@ -35,7 +36,7 @@ var uninstallCmdDeprecated = &cobra.Command{
 	Short:      uninstallCmdShortDesc,
 	Long:       uninstallCmdLongDesc,
 	Example:    uninstallCmdExamples,
-	Deprecated: "instead use \"" + cmd.K8sCmdLiteral + " " + cmd.K8sUninstallCmdLiteral + "\".",
+	Deprecated: "instead use \"" + k8s.K8sCmdLiteral + " " + k8s.K8sUninstallCmdLiteral + "\".",
 }
 
 func init() {
