+++ import-export-cli/cmd/k8s/get.go
@@ -1,0 +1,1 @@
+package k8s
