+++ import-export-cli/cmd/k8s/getApi.go
@@ -1,0 +1,1 @@
+package k8s
