+++ import-export-cli/cmd/k8s/k8s.go
@@ -1,25 +1,8 @@
-/*
-*  Copyright (c) WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
- */
-
-package cmd
+package k8s
 
 import (
 	"github.com/spf13/cobra"
+	cmd2 "github.com/wso2/product-apim-tooling/import-export-cli/cmd"
 	"github.com/wso2/product-apim-tooling/import-export-cli/utils"
 )
 
@@ -42,12 +25,13 @@ var K8sCmd = &cobra.Command{
 	Long:    k8sCmdLongDesc,
 	Example: k8sCmdExamples,
 	Run: func(cmd *cobra.Command, args []string) {
-		utils.Logln(utils.LogPrefixInfo + ImportCmdLiteral + " called")
+		utils.Logln(utils.LogPrefixInfo + cmd2.ImportCmdLiteral + " called")
 
 	},
 }
 
 // init using Cobra
 func init() {
-	RootCmd.AddCommand(K8sCmd)
+	cmd2.RootCmd.AddCommand(K8sCmd)
 }
+
