+++ import-export-cli/cmd/k8s/uninstall.go
@@ -1,22 +1,4 @@
-/*
-*  Copyright (c) WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
- */
-
-package cmd
+package k8s
 
 import (
 	"github.com/spf13/cobra"
@@ -39,3 +21,4 @@ var uninstallCmd = &cobra.Command{
 func init() {
 	K8sCmd.AddCommand(uninstallCmd)
 }
+
