+++ import-export-cli/go.mod
@@ -1,11 +1,11 @@
 go 1.12
 
 require (
-	github.com/Jeffail/gabs v1.2.0
+	github.com/Jeffail/gabs v1.4.0
 	github.com/cpuguy83/go-md2man v1.0.10 // indirect
 	github.com/getkin/kin-openapi v0.2.0
 	github.com/ghodss/yaml v1.0.1-0.20190212211648-25d852aebe32
-	github.com/go-openapi/loads v0.19.4
+	github.com/go-openapi/loads v0.19.5
 	github.com/go-resty/resty v0.0.0-20171018191538-8b5e3f91fbea
 	github.com/google/go-cmp v0.4.0
 	github.com/google/uuid v1.1.1
@@ -13,15 +13,15 @@ require (
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/json-iterator/go v1.1.10
 	github.com/magiconair/properties v1.8.1
-	github.com/mitchellh/go-homedir v1.1.0 // indirect
-	github.com/mitchellh/mapstructure v1.1.2
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/mitchellh/mapstructure v1.3.2
 	github.com/renstrom/dedent v1.0.0
-	github.com/spf13/cast v1.3.0
+	github.com/spf13/cast v1.3.1
 	github.com/spf13/cobra v1.1.1
-	github.com/stretchr/testify v1.4.0
+	github.com/stretchr/testify v1.6.1
 	github.com/wso2/k8s-api-operator/api-operator v0.0.0-20200722085006-f178a61e4fde
-	golang.org/x/crypto v0.0.0-20200220183623-bac4c82f6975
-	gopkg.in/yaml.v2 v2.2.8
+	golang.org/x/crypto v0.0.0-20200414173820-0848c9571904
+	gopkg.in/yaml.v2 v2.3.0
 	sigs.k8s.io/testing_frameworks v0.1.1 // indirect
 	sigs.k8s.io/yaml v1.2.0 // indirect
 	github.com/pavel-v-chernykh/keystore-go/v4 v4.1.0
